refactor: ChickenTable 연결 & 데모페이지 삭제
