+++ src/components/ChickenTable/ChickenTable.tsx
@@ -1,21 +1,29 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../../store";
-import { setCurrentPage, setData, toggleSelect } from "../../store/slices/chickenSlice";
+import { AppDispatch, RootState } from "../../store";
+import { setCurrentPage, setData } from "../../store/slices/chickenSlice";
 import { ChickenData } from "../../types/ChickenData";
 import "./ChickenTable.css";
+import { setDetailData } from "../../store/slices/detailSlice";
 
 interface ChickenTableProps {
     filteredData: ChickenData[];
 }
 
 const ChickenTable: React.FC<ChickenTableProps> = ({ filteredData }) => {
-    const dispatch = useDispatch();
+    const dispatch = useDispatch<AppDispatch>();
     const data = useSelector((state: RootState) => state.chicken.data);
-    const selected = useSelector((state: RootState) => state.chicken.selected);
     const currentPage = useSelector((state: RootState) => state.chicken.currentPage);
     const itemsPerPage = 15;
     const pageNumbersPerPage = 10;
+
+    // 체크박스 선택
+    const detailOnSelectedItems = useSelector(
+      (state: RootState) => state.detail
+    );
+    const handleOnCheckboxChange = (item: ChickenData) => {
+      dispatch(setDetailData(item));
+    };    
 
     // 데이터 가져오기
     useEffect(() => {
@@ -23,10 +31,6 @@ const ChickenTable: React.FC<ChickenTableProps> = ({ filteredData }) => {
             .then((response) => response.json())
             .then((data) => dispatch(setData(data)));
     }, [dispatch]);
-
-    const handleCheckboxChange = (item: ChickenData) => {
-        dispatch(toggleSelect(item));
-      };
 
     // 테이블 페이징
     const indexOfLastItem = currentPage * itemsPerPage;
@@ -55,59 +59,57 @@ const ChickenTable: React.FC<ChickenTableProps> = ({ filteredData }) => {
 
 
     return (
-        <div className="table-container">
-            <table className="chicken-table">
-                <thead>
-                    <tr>
-                        <th>선택</th>
-                        <th>브랜드</th>
-                        <th>제품명</th>
-                        <th>단백질(g)</th>
-                        <th>에너지(kcal)</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {currentItems.map((item) => (
-                        <tr key={item.id}>
-                            <td>
-                                <input
-                                    type="checkbox"
-                                    checked={
-                                        !!selected.find((i) => i.id === item.id)
-                                    }onChange={() => handleCheckboxChange(item)}
-                                    
-                                />
-                            </td>
-                            <td>{item.brand}</td>
-                            <td>{item.product_name}</td>
-                            <td>{item.protein}</td>
-                            <td>{item.calories}</td>
-                        </tr>
-                    ))}
-                </tbody>
-            </table>
-            <div className="pagination">
-                {startPage > 1 && (
-                    <button onClick={() => paginate(startPage - 1)}>
-                        &laquo;
-                    </button>
-                )}
-                {pageNumbers.slice(startPage - 1, endPage).map((number) => (
-                    <button
-                        key={number}
-                        onClick={() => paginate(number)}
-                        className={currentPage === number ? "active" : ""}
-                    >
-                        {number}
-                    </button>
-                ))}
-                {endPage < totalPageNumbers && (
-                    <button onClick={() => paginate(endPage + 1)}>
-                        &raquo;
-                    </button>
-                )}
-            </div>
+      <div className="table-container">
+        <table className="chicken-table">
+          <thead>
+            <tr>
+              <th>선택</th>
+              <th>브랜드</th>
+              <th>제품명</th>
+              <th>단백질(g)</th>
+              <th>에너지(kcal)</th>
+            </tr>
+          </thead>
+          <tbody>
+            {currentItems.map((item) => (
+              <tr key={item.id}>
+                <td>
+                  <input
+                    type="checkbox"
+                    checked={
+                      !!detailOnSelectedItems.selectedData.find(
+                        (i) => i.id === item.id
+                      )
+                    }
+                    onChange={() => handleOnCheckboxChange(item)}
+                  />
+                </td>
+                <td>{item.brand}</td>
+                <td>{item.product_name}</td>
+                <td>{item.protein}</td>
+                <td>{item.calories}</td>
+              </tr>
+            ))}
+          </tbody>
+        </table>
+        <div className="pagination">
+          {startPage > 1 && (
+            <button onClick={() => paginate(startPage - 1)}>&laquo;</button>
+          )}
+          {pageNumbers.slice(startPage - 1, endPage).map((number) => (
+            <button
+              key={number}
+              onClick={() => paginate(number)}
+              className={currentPage === number ? "active" : ""}
+            >
+              {number}
+            </button>
+          ))}
+          {endPage < totalPageNumbers && (
+            <button onClick={() => paginate(endPage + 1)}>&raquo;</button>
+          )}
         </div>
+      </div>
     );
 };
